refactor and add duration
